+++ CHANGELOG.md
@@ -50,6 +50,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * `jj hide` (alias for `jj abandon`) is no longer available. Use `jj abandon`
   instead.
 
+* `jj debug completion`, `jj debug mangen` and `jj debug config-schema` have
+  been moved from `jj debug` to `jj support`.
+
 ### Fixed bugs
 
 * Modify/delete conflicts now include context lines